screenshot: extend the filter for startup images
