+++ _includes/card-job.html
@@ -63,7 +63,7 @@ Ex : {% include card-job.html job = job %}
   </div>
   <div class="fr-card__header">
     <div class="fr-card__img">
-      {% include screenshot.html startup=startup force=true %}
+      {% include screenshot.html startup=startup %}
     </div>
   </div>
 </div>
