+++ _includes/card-startup.html
@@ -20,7 +20,7 @@
       <p class="fr-card__desc">{{ startup.mission | strip_html }}</p>
     </div>
     <div class="fr-card__img">
-      {% include screenshot.html startup=startup force=true %}
+      {% include screenshot.html startup=startup %}
     </div>
   </div>
 </div>
