+++ _layouts/home.html
@@ -55,7 +55,7 @@ layout: default
                   <p class="fr-card__desc">{{ startup.mission | strip_html }}</p>
                 </div>
                 <div class="fr-card__img">
-                  {% include screenshot.html startup=startup force=true %}
+                  {% include screenshot.html startup=startup %}
                 </div>
               </div>
             </div>
